core: move rag pipeline implementation to runtime
