+++ src/runtime/generation/__init__.py
@@ -1,0 +1,6 @@
+"""Runtime generation exports."""
+
+from .contracts import GenerationCallable, GenerationRequest, GenerationResult
+from .service import generate_answer
+
+__all__ = ["GenerationCallable", "GenerationRequest", "GenerationResult", "generate_answer"]
