+++ telegram_bot/pipelines/client.py
@@ -12,7 +12,8 @@ from typing import Any
 
 from src.observability_payloads import build_safe_input_payload
 from src.retrieval.topic_classifier import get_query_topic_hint
-from telegram_bot.agents.rag_pipeline import rag_pipeline
+from src.runtime.grounding.policy import get_grounding_mode
+from src.runtime.pipeline.rag import rag_pipeline
 from telegram_bot.graph.nodes.respond import _MAX_SOURCES, format_sources
 from telegram_bot.observability import get_client, observe, propagate_attributes
 from telegram_bot.pipelines.state_contract import coerce_pre_agent_state_contract
@@ -25,7 +26,6 @@ from telegram_bot.services.cache_policy import (
     resolve_semantic_cache_signature,
 )
 from telegram_bot.services.generate_response import generate_response
-from telegram_bot.services.grounding_policy import get_grounding_mode
 from telegram_bot.services.history_service import HistoryService
 from telegram_bot.services.telegram_formatting import send_html_messages
 from telegram_bot.services.types import PipelineResult
